+++ react/components/AffirmModal.tsx
@@ -1,140 +1,189 @@
 /* eslint-disable @typescript-eslint/camelcase */
-import React, { useEffect } from 'react'
+import React, { Component } from 'react'
 import Helmet from 'react-helmet'
-import { useQuery, useMutation, MutationFetchResult } from 'react-apollo'
-import useScriptLoader from './hooks/useScriptLoader'
-import { useRuntime, canUseDOM } from 'vtex.render-runtime'
-import Settings from '../graphql/Settings.graphql'
-import OrderData from '../graphql/OrderData.graphql'
+import { graphql, MutationOptions, FetchResult } from 'react-apollo'
+import withSettings from './withSettings'
+import withOrderData from './withOrderData'
 import OrderUpdate from './../graphql/OrderUpdate.graphql'
 
-interface AffirmAuthorizationProps {
+interface AffirmAuthenticationProps {
   appPayload: string
+  settings: AffirmSettings
+  orderData: OrderData
+  orderUpdateMutation(options?: MutationOptions): Promise<FetchResult>
+}
+
+interface AffirmSettings {
+  isLive: boolean
+  enableKatapult: boolean
+  companyName: string
+}
+
+interface OrderData {
+  transactionId: string
+  orderId: string
+  miniCart: any
+  value: number
+  inboundRequestsUrl: string
+  callbackUrl: string
 }
 
 declare const vtex: any
 declare const $: any
 
-const AffirmModal: StorefrontFunctionComponent<AffirmAuthorizationProps> = ({
-  appPayload,
-}) => {
-  const { rootPath } = useRuntime()
-  const { data: settingsData } = useQuery(Settings, { ssr: false })
-  const { paymentIdentifier, publicKey } = JSON.parse(appPayload)
-  const { data: orderQueryResult } = useQuery(OrderData, {
-    skip: !paymentIdentifier,
-    variables: {
-      qs: paymentIdentifier,
-    },
-    ssr: false,
-  })
-  const siteRoot = rootPath || ''
-  const [updateOrder] = useMutation(OrderUpdate)
+class AffirmModal extends Component<AffirmAuthenticationProps> {
+  public state = {
+    scriptLoaded: false,
+  }
 
-  const [affirm, { error }] = useScriptLoader(
-    settingsData?.affirmSettings?.isLive
-      ? 'https://cdn1.affirm.com/js/v2/affirm.js'
-      : 'https://cdn1-sandbox.affirm.com/js/v2/affirm.js',
-    'affirm'
-  )
+  // eslint-disable-next-line react/no-deprecated
+  public componentWillMount = () => {
+    this.inject()
+    window.modalTriggered = false
+  }
 
-  const respondTransaction = (status: boolean) => {
+  public async inject() {
+    const {
+      settings: { isLive },
+    } = this.props
+
+    if (isLive == null) {
+      console.error('No Affirm settings found')
+      await this.respondTransaction(false)
+    }
+
+    const affirmUrl = isLive
+      ? `https://cdn1.affirm.com/js/v2/affirm`
+      : `https://cdn1-sandbox.affirm.com/js/v2/affirm`
+
+    this.injectScript(
+      'affirm-checkout-sdk-js',
+      `${affirmUrl}.js`,
+      this.handleOnLoad
+    )
+  }
+
+  public affirm: any = {}
+
+  public async respondTransaction(status: boolean) {
+    if (!status) {
+      await fetch(`${this.props.orderData.callbackUrl}?status=denied`, {
+        method: 'post',
+        mode: 'no-cors',
+      })
+    }
     $(window).trigger('transactionValidation.vtex', [status])
   }
 
-  useEffect(() => {
-    window.modalTriggered = false
-  }, [])
+  public injectScript = (id: string, src: string, onLoad: () => void) => {
+    if (document.getElementById(id)) {
+      return
+    }
 
-  useEffect(() => {
-    if (
-      orderQueryResult?.orderData &&
-      settingsData?.affirmSettings &&
-      affirm &&
-      !error &&
-      !window.modalTriggered &&
-      publicKey
-    ) {
-      const { orderData } = orderQueryResult
+    // eslint-disable-next-line prefer-destructuring
+    const head = document.getElementsByTagName('head')[0]
+
+    const js = document.createElement('script')
+    js.id = id
+    js.src = src
+    js.async = true
+    js.defer = true
+    js.onload = onLoad
+
+    head.appendChild(js)
+  }
+
+  public handleOnLoad = () => {
+    this.setState({ scriptLoaded: true })
+  }
+
+  public componentDidUpdate() {
+    const {
+      orderData,
+      orderData: { miniCart },
+      settings: { companyName, enableKatapult },
+      orderUpdateMutation,
+    } = this.props
+    const { scriptLoaded } = this.state
+
+    const { publicKey } = JSON.parse(this.props.appPayload)
+    if (scriptLoaded && !window.modalTriggered) {
       window.modalTriggered = true
       window._affirm_config = {
         public_api_key: publicKey,
       }
-      const items = orderQueryResult.orderData.miniCart.items.map(
-        (item: any) => ({
-          display_name: item.name,
-          sku: item.id,
-          unit_price: item.price * 100,
-          qty: item.quantity,
-          leasable: true,
-        })
-      )
-
-      vtex.checkout.MessageUtils.hidePaymentMessage()
-
-      affirm.checkout({
+      const affirmItems = orderData.miniCart.items.map((item: any) => ({
+        leasable: true,
+        display_name: item.name,
+        sku: item.id,
+        unit_price: item.price * 100,
+        qty: item.quantity,
+      }))
+      window.affirm.checkout({
         merchant: {
-          exchange_lease_enabled:
-            settingsData?.affirmSettings?.enableKatapult ?? false,
-          name: settingsData?.affirmSettings?.companyName ?? '',
+          exchange_lease_enabled: enableKatapult,
+          name:
+            companyName != null && companyName != '' ? companyName : undefined,
           user_confirmation_url: '',
           user_cancel_url: '',
           user_confirmation_url_action: 'GET',
         },
         shipping: {
           name: {
-            first: orderData.miniCart.buyer.firstName,
-            last: orderData.miniCart.buyer.lastName,
+            first: miniCart.buyer.firstName,
+            last: miniCart.buyer.lastName,
           },
           address: {
-            line1: orderData.miniCart.shippingAddress.street,
-            line2: orderData.miniCart.billingAddress.complement,
-            city: orderData.miniCart.shippingAddress.city,
-            state: orderData.miniCart.shippingAddress.state,
-            zipcode: orderData.miniCart.shippingAddress.postalCode,
-            country: orderData.miniCart.shippingAddress.country,
+            line1: miniCart.shippingAddress.street,
+            line2: miniCart.shippingAddress.complement,
+            city: miniCart.shippingAddress.city,
+            state: miniCart.shippingAddress.state,
+            zipcode: miniCart.shippingAddress.postalCode,
+            country: miniCart.shippingAddress.country,
           },
-          phone_number: orderData.miniCart.buyer.phone,
-          email: orderData.miniCart.buyer.email,
+          phone_number: miniCart.buyer.phone,
+          email: miniCart.buyer.email,
         },
         billing: {
           name: {
-            first: orderData.miniCart.buyer.firstName,
-            last: orderData.miniCart.buyer.lastName,
+            first: miniCart.buyer.firstName,
+            last: miniCart.buyer.lastName,
           },
           address: {
-            line1: orderData.miniCart.billingAddress.street,
-            line2: orderData.miniCart.billingAddress.complement,
-            city: orderData.miniCart.billingAddress.city,
-            state: orderData.miniCart.billingAddress.state,
-            zipcode: orderData.miniCart.billingAddress.postalCode,
-            country: orderData.miniCart.billingAddress.country,
+            line1: miniCart.billingAddress.street,
+            line2: miniCart.billingAddress.complement,
+            city: miniCart.billingAddress.city,
+            state: miniCart.billingAddress.state,
+            zipcode: miniCart.billingAddress.postalCode,
+            country: miniCart.billingAddress.country,
           },
-          phone_number: orderData.miniCart.buyer.phone,
-          email: orderData.miniCart.buyer.email,
+          phone_number: miniCart.buyer.phone,
+          email: miniCart.buyer.email,
         },
-        items: items,
+        items: affirmItems,
         metadata: {
           shipping_type: '',
           mode: 'modal',
         },
         order_id: orderData.orderId,
-        shipping_amount: orderData.miniCart.shippingValue * 100,
-        tax_amount: orderData.miniCart.taxValue * 100,
+        shipping_amount: miniCart.shippingValue * 100,
+        tax_amount: miniCart.taxValue * 100,
         total: orderData.value * 100,
       })
-      affirm.checkout.open({
+      vtex.checkout.MessageUtils.hidePaymentMessage()
+      var self = this
+      window.affirm.checkout.open({
         // eslint-disable-next-line prettier/prettier
-        onFail: function () {
+        onFail: async function () {
           vtex.checkout.MessageUtils.showPaymentMessage()
-          respondTransaction(false)
+          await self.respondTransaction(false)
         },
         // eslint-disable-next-line prettier/prettier
-        onSuccess: function (a: any) {
+        onSuccess: async function (a: any) {
           vtex.checkout.MessageUtils.showPaymentMessage()
-          updateOrder({
+          const response = await orderUpdateMutation({
             variables: {
-              url: orderData.orderData.inboundRequestsUrl
+              url: orderData.inboundRequestsUrl
                 .replace('https', 'http')
                 .replace(':action', 'auth'),
               orderId: orderData.orderId,
@@ -142,56 +191,43 @@ const AffirmModal: StorefrontFunctionComponent<AffirmAuthorizationProps> = ({
               callbackUrl: orderData.callbackUrl,
               orderTotal: orderData.value * 100,
             },
-          }).then((response: MutationFetchResult) => {
-            if (response.data && response.data.orderUpdate) {
-              respondTransaction(true)
-            } else {
-              respondTransaction(false)
-            }
           })
+          if (response.data && response.data.orderUpdate) {
+            await self.respondTransaction(true)
+          } else {
+            await self.respondTransaction(false)
+          }
         },
       })
-      // eslint-disable-next-line prettier/prettier
-      affirm.ui.error.on('close', function () {
+
+      window.affirm.ui.error.on('close', async function () {
         vtex.checkout.MessageUtils.showPaymentMessage()
-        respondTransaction(false)
+        await self.respondTransaction(false)
       })
     }
-  }, [
-    affirm,
-    error,
-    orderQueryResult,
-    publicKey,
-    settingsData.affirmSettings,
-    siteRoot,
-    updateOrder,
-  ])
-
-  if (
-    !canUseDOM ||
-    !settingsData?.affirmSettings ||
-    !orderQueryResult?.orderData ||
-    !publicKey ||
-    !affirm
-  )
-    return null
-
-  if (error) {
-    respondTransaction(false)
-    return null
   }
 
-  return (
-    <Helmet>
-      <script>
-        {`
-            _affirm_config = {
-            public_api_key:  "${publicKey}"
-            };
-        `}
-      </script>
-    </Helmet>
-  )
+  public render() {
+    const { scriptLoaded } = this.state
+    const { publicKey } = JSON.parse(this.props.appPayload)
+
+    return (
+      scriptLoaded &&
+      publicKey && (
+        <Helmet>
+          <script>
+            {`
+              _affirm_config = {
+              public_api_key:  "${publicKey}"
+              };
+          `}
+          </script>
+        </Helmet>
+      )
+    )
+  }
 }
 
-export default AffirmModal
+export default graphql(OrderUpdate, { name: 'orderUpdateMutation' })(
+  withSettings(withOrderData(AffirmModal))
+)