+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Changed to class component to meet requirements of react@2.x builder
+
 ## [2.0.1] - 2020-04-03
 
 ### Fixed